Better performance of sbml file patching
Previously, certain patches to the output SBML file were performed using
perl, which was called from within a R session. This solution was
working correctly, but caused long run times (very long on some
machines).

With this commit, no perl is called and all patches are done with R
only.

Caution: SBML patching might need to be updated with updates in libsbml
and/or sybilSBML package.