Patching gene product IDs in SBML output files
Within R, genes have IDs in the following syntax
<ContigID>_<startPos>:<endPos>

This syntax makes it easy to locate the position of genes that are
associated with reactions or transporters.

libsbml in combination with sybilSBML has caused that certain characters
were reformed for sbml export. E.g.: "123" -> "__ONE____TWO____THREE__".

With this commit, the gene names/gene product names are back formated to
the orignal names. Please note, that all ".","-", or ":" are replaced
with an underscore "_" in the sbml file.