Minor bug fix (empty models)
An error in the SBML export occured, if the now reactions are added in
the draft network reconstruction, which could happen if a too short or
random DNA sequence was used as input to gapseq find and find-transport.

Fix: In these rare cased, include a dummy reaction to the model: ATP
phosphohydrolase.

In these cases, the models should not be used for any subsequent
predictions / analysis / interpretations.