Speed up gapseq find by reusing prev. blast results for same query
Before: When different reference protein fasta files were combined (e.g.
sequence set fro swissprot and metacyc) a new temporary fasta is
created.

The problem: gapseq checks whether a blast query was performed before,
but this worked not if new temporary fasta with a random name was
created.

The Solution: Saving the blast results with a name that is the md5
checksum of the newly combined fasta file. In this way, gapseq can
recognised if the same sequnces were already aligned to the input
genome.

The challenge: Complex handling. But this should be resolved now. See
new lines 801 and 830-832.

Pathway that illustrated the speed-up: PWYG-321