Speed up by restructuring complex detection script
Complex detection is done in an R-script, that is called from within
"gapseq find". Until know, this rscript loaded the R-package
"Biostrings" as dependency. Loading Biostrings requires 1-3 seconds to
load. No much but when the rscript is called >1000 times in "gapseq
find", it significantly contributes to overall run time.

Now: Removed dependency on "Biostrings".