Add openssl headers to install instructions
On a clean Ubuntu 22.04 install, R packages fail to install due to dependency on R package openssl, which fails to build. Installing the package libssl-dev fixes this. I believe that the corresponding package for Fedora/CentOS might be called openssl-devel, but can't test it (info from https://github.com/jeroen/openssl#installation). It also seems like this is not needed on Mac.