Fix #149
Special characters in sequence titles causes errors in GPR rules in
model object.

Furthermore, very long gene names (which are derived from long sequence
titles of the fasta input file) caused a "buffer overfow" error in the
SBML export using libSBML.

In this fix, special characters are replaced with underscores and long
sequence identifiers are truncated to a max length of 40 characters. It
is automatically tested if sequence identifiers remain unique after the
adjustments; if not, new unique generic sequence identifiers are
assigned to sequences.