Better gene naming when genome input is protein fasta
When a genomic nucleotide fasta is provided as input, gapseq assigns
gene names in the form of

<contigID>_<start>:<end>

This naming does not make sense when the input genome is a protein
fasta. When a protein fasta is provided, genes are simply named in the
form:

<protinID>

Where the protein ID is obtained from the fasta sequence titles.

Please note: A new header comment is added to the gapseq output files
"...-Pathways.tbl" and "...-Reactions.tbl", which indicates the input
genome format ["prot"/"nucl"].