Enabled tRNA-dependent asparagine biosynthesis
The pathway (metacyc: PWY490-4) was block by missing reactions.

Also, please note the the new reacton rxn90115 (EC: 6.3.5.6) is not 100
percent correct, because it forms free L-asparagine. I implemented it
this way, because the biomass reactions in gapseq (an in most
genome-scale models) free asparagine flows into the biomass. Asn-charged
tRNA currently is not part of biomass reactions used in gapseq.

Propectively, we could automatically adjust the biomass reaction in
cases where the tRNA-dependent biosynthesis is the only pathway for
Asn-biosynthesis and replace free Asn with the Asn-charged tRNA in the
biomass definition.