Bug fix in GPR Associations
In case no genes are associated to the first reaction added to the model,
no 'genes','grp' and 'gprRules' objects are initiated by sybil, causing a
mismatch between length of the number of reactions and those three objects
in the final modelorg model. This mismatch causes libSBML to stop on a
segmentation error. Solution: "Manually" initiate the three objects in those
cases.

This bug was rare (less than 1% of the genomes tested from the HRGM
collection)