Sequence-DB error fix
It seems that in the metacyc genes there is a mis-assignment to Uniprot
for the reaction: SULFATE-ADENYLYLTRANS-RXN

Instead of Q92R20 it should be P56893 .

I contacted MetaCyc and suggested the change in their DB.

In the meanwhile, the patched sequence file overwrites the sequence file
from metacyc. As soon as MetaCyc has fixed the mis-match and we have
updated dat/meta_genes.csv, the file dat/seq/Bacteria/user/SULFATE-ADENYLYLTRANS-RXN.fasta can be deleted.

closes #165