Bug fix in enumerating candidate reactions linked to complexes
The "...-all-Reactions.tbl" output table should list all subunits of
protein complexes, even if they are not found by sequence alignments to
reference proteins. When implementing the batch alignment approach, we
accidently lost this behavious, which let to missing 'no_blast' entries
in the output table.

Listing all subunits in the "...-all-Reactions.tbl" is important,
because the info on subunits with "no_blast" status inform the calculations of
reactions scores (during `gapseq draft`) used for later gapfilling.

The bug had the effect, that we overpredicted reactions associated to
protein complex enzymes, especially in those cases, where most subunits
were not found but only one or very few with bitscores > 100.

I noticed the bug with the enzyme EC 1.1.1.436 (electron confurcating
LDH).