Small fix in pathway duplicate filtering
When duplicate pathways are identified in the gapseq find workflow,
gapseq will use only the pathway definition from the custom database. In
the previous filtering of the pathways there was the potential issue
that the pathway 'NADPHOS-DEPHOS-PWY' matched also the pathway ID
'NADPHOS-DEPHOS-PWY-1' , which is not desired because these are separate
pathways.

The new, awk-based, implementation makes sure, that only duplicates with
exact same IDs (from start to end) are filtered.