Added skatole production reaction and pathway
Skatole can be a product of microbial tryptophan degradation, e.g., in
animal gut microbiomes.

The reaction with the EC 4.1.1.115 was absent in the biochemistry DB and
not associated with any pathway.

Reaction was added to the DB (rxn90146) and a skatole diffusion reaction
(rxn90147).

Also, a one-reaction pathway for skatole production from indoleacetate
was added to the custom pathway table.